feat(socketio/extract): refactor extract mod
